INTERNAL: do_htree_traverse_sampling/elem_at_offset를 hash_tree 모듈로 추출
- do_htree_traverse_sampling: reservoir sampling 순회를 hash_tree로 추출
- do_htree_elem_at_offset: offset 기반 elem 탐색/삭제를 hash_tree로 추출, on_delete 콜백 지원
- do_set_elem_traverse_sampling, do_set_elem_at_offset 제거
- do_set_elem_unlink: 호출 지점 없어져 제거
- hash_tree.c: <stdlib.h> include 추가 (rand 사용)
